move hyperdrive specific functions
